requirements.txt
aiohttp
aiologger
python-socketio
requests